restructure observation-api
