+++ src/libs/observable-api/states/boolean-state.ts
