+++ src/libs/observable-api/states/number-state.ts
