+++ src/libs/observable-api/states/string-state.ts
