+++ src/libs/observable-api/types/index.ts
@@ -1,0 +1,2 @@
+export * from './mapping-function.type.js';
+export * from './memoization-function.type.js';
