+++ src/libs/observable-api/types/mapping-function.type.ts
