+++ src/libs/observable-api/utils/deep-freeze.function.ts
