+++ src/libs/observable-api/utils/default-memoization.function.ts
@@ -1,4 +1,4 @@
-import { naiveObjectComparison } from './naive-object-comparison.js';
+import { naiveObjectComparison } from './naive-object-comparison.function.js';
 
 export function defaultMemoization(previousValue: any, currentValue: any): boolean {
 	if (typeof previousValue === 'object' && typeof currentValue === 'object') {
