+++ src/libs/observable-api/utils/filter-frozen-array.function.ts
