+++ src/libs/observable-api/utils/partial-update-frozen-array.function.ts
